Refactor TUI input handling loop
🚮 Smell: The TUI input loop (`handle_input`) contained deeply nested logic (`match` inside `if let` inside `match`), making `tui.rs` very hard to read and modifying input modes bug-prone.
✨ Solution: Extracted the input loop components into named helper functions (`handle_spirit_request`, `handle_injection_mode`, `handle_editing_mode`, `handle_normal_mode`) returning `Result<bool>`.
🧼 Benefit: Significantly flattened the structure, improving readability without changing logic.
🛡️ Verification: Tests passed. No logic was altered; the application state mutations correctly mirror the old implementation. Also ran `cargo clippy --fix` on the entire repo to tidy up warning noise across several nova features.

+++ experiments/chimera-lang/src/vm/nova_pocket.rs
@@ -12,7 +12,7 @@ pub fn exec_pocket(vm: &mut ChimeraVM) -> Option<(usize, usize)> {
     if let Some(val) = vm.stack.pop() {
         if let Value::Int(r) = val {
             if r >= 0 {
-                let radius = r as i64;
+                let radius = r;
                 let (cy, cx) = vm.context_loc;
                 let mut data = Vec::new();
